Programs

+++ LeapYear.java
@@ -6,7 +6,7 @@ public class LeapYear
     {
         Scanner scan=new Scanner(System.in);
         System.out.println("Enter the year: ");
-        long year= scan.nextLong();
+        int year= scan.nextInt();
         if((year%4==0) && (year%100!=0) ||(year%400==0))
         {
             System.out.println(year+" is a Leap Year");
